Moved store update on layer change from DeckGLMap to Map
